misc: Major repo cleanup
Fix a lot of technical debt and complexity in package.json.

Change-type: patch
