--- lib/action-library/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@resin.io/jellyfish-action-library",
-  "private": true,
-  "main": "dist/index.js",
-  "author": "Resin Inc. <hello@resin.io>",
-  "contributors": [
-    "Andrew Lucas <andrewl@resin.io>"
-  ],
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@types/node": "^10.7.1"
-  }
-}
