+++ lib/action-library/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "../../tsconfig",
   "compilerOptions": {
     "outDir": "./dist",
-    "rootDir": "./lib"
+    "rootDir": "."
   },
   "include": [
     "**/*.ts"
