+++ lib/core/backend.js
@@ -20,7 +20,7 @@ const EventEmitter = require('events').EventEmitter
 const debug = require('debug')('jellyfish:backend')
 const Bluebird = require('bluebird')
 const reqlSchema = require('reql-schema')
-const jsonSchema = require('../json-schema')
+const skhema = require('skhema')
 const jellyscript = require('../jellyscript')
 const errors = require('./errors')
 
@@ -159,8 +159,8 @@ const createTableStream = async (backend, table, schema) => {
 			return
 		}
 
-		const newMatch = jsonSchema.filter(schema, change.new_val)
-		const oldMatch = jsonSchema.filter(schema, change.old_val)
+		const newMatch = skhema.filter(schema, change.new_val)
+		const oldMatch = skhema.filter(schema, change.old_val)
 
 		if (newMatch || oldMatch) {
 			emitter.emit('data', {
@@ -201,7 +201,7 @@ const queryTable = async (backend, table, schema, options) => {
 	// using JSON schema. This is meant to be used as a prototype,
 	// and we should implement a proper JSON Schema to ReQL translator.
 	const items = await cursor.toArray()
-	return jsonSchema.filter(schema, items)
+	return skhema.filter(schema, items)
 }
 
 const getElementByIdFromTable = async (backend, table, id) => {
@@ -828,7 +828,7 @@ module.exports = class Backend {
 				return []
 			}
 
-			return _.compact([ jsonSchema.filter(schema, _.cloneDeep(element)) ])
+			return _.compact([ skhema.filter(schema, _.cloneDeep(element)) ])
 		}
 
 		const sortExpression = schema[SORT_KEY]
