+++ lib/core/kernel.js
@@ -15,7 +15,7 @@
  */
 
 const debug = require('debug')('jellyfish:kernel')
-const jsonSchema = require('../json-schema')
+const jsonSchema = require('./json-schema')
 const errors = require('./errors')
 const CARDS = require('./cards')
 const permissionFilter = require('./permission-filter')
