+++ lib/core/permission-filter.js
@@ -16,7 +16,7 @@
 
 const _ = require('lodash')
 const jsone = require('json-e')
-const jsonSchema = require('../json-schema')
+const jsonSchema = require('./json-schema')
 const errors = require('./errors')
 const CARDS = require('./cards')
