--- lib/index.js
@@ -1,3 +1,0 @@
-const createServer = require('./server')
-
-createServer()
