--- lib/sdk/.editorconfig
@@ -1,13 +1,0 @@
-# editorconfig.org
-root = true
-
-[*]
-indent_style = tab
-indent_size = 2
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-
-[*.{yml,yaml,json}]
-indent_style = space
