+++ lib/server/action-server.js
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-const Worker = require('./worker')
-const actionLibrary = require('./action-library')
+const Worker = require('../worker')
+const actionLibrary = require('../action-library')
 
 const SCHEMA_ACTIVE_TRIGGERS = {
 	type: 'object',
