+++ lib/server/error-reporter.js
