+++ lib/server/routes.js
@@ -23,7 +23,7 @@ const uuid = require('uuid/v4')
 const {
 	version
 } = require('../../package')
-const Storage = require('../file-storage')
+const Storage = require('./file-storage')
 
 const fileStore = new Storage({
 	driver: process.env.FS_DRIVER || 'localFS'
