+++ lib/worker/executor.js
@@ -16,7 +16,7 @@
 
 const Bluebird = require('bluebird')
 const _ = require('lodash')
-const jsonSchema = require('../json-schema')
+const skhema = require('skhema')
 const utils = require('./utils')
 const errors = require('./errors')
 const triggers = require('./triggers')
@@ -267,12 +267,12 @@ exports.run = async (jellyfish, session, context, library, request) => {
 		type: 'object'
 	})
 
-	if (!jsonSchema.isValid(actionInputCardFilter, cards.input)) {
+	if (!skhema.isValid(actionInputCardFilter, cards.input)) {
 		throw new errors.WorkerSchemaMismatch('Input card does not match filter')
 	}
 
 	const argumentsSchema = utils.getActionArgumentsSchema(request.action)
-	if (!jsonSchema.isValid(argumentsSchema, request.arguments)) {
+	if (!skhema.isValid(argumentsSchema, request.arguments)) {
 		throw new errors.WorkerSchemaMismatch('Arguments do not match')
 	}
