+++ lib/worker/triggers.js
@@ -15,7 +15,7 @@
  */
 
 const jsone = require('json-e')
-const jsonSchema = require('../json-schema')
+const skhema = require('skhema')
 
 /**
  * @summary Check if a trigger matches a card
@@ -44,7 +44,7 @@ const matchesCard = (trigger, card) => {
 		return false
 	}
 
-	return jsonSchema.isValid(trigger.filter, card)
+	return skhema.isValid(trigger.filter, card)
 }
 
 /**
