+++ package.json
@@ -10,24 +10,11 @@
     "url": "git@github.com:resin-io/jellyfish.git"
   },
   "scripts": {
-    "start": "DEBUG=jellyfish:* node lib/index.js",
-    "start:profiler": "DEBUG=jellyfish:* node profiler.js",
-    "test": "npm run lint && npm run test:command",
-    "test:command": "TEST_DB_HOST=localhost TEST_PROXY_PORT=9999 TEST_DB_PORT=28015 API_URL='http://localhost:8000' NODE_ENV=test nyc --reporter=lcov ava",
+    "start": "make start",
+    "test": "make lint test",
     "test:e2e": "ava e2e",
-    "posttest": "./scripts/scrub-test-databases.js",
-    "lint": "npm run eslint && npm run tslint",
-    "eslint": "eslint lib test stress webpack.config.js",
-    "eslint:fix": "eslint --fix lib test webpack.config.js",
-    "tslint": "tslint -t stylish \"lib/**/*.ts\" \"lib/**/*.tsx\" -e \"lib/*/node_modules/**\" -e \"lib/*/dist/**\"",
-    "tslint:fix": "tslint -t stylish --fix \"lib/**/*.ts\" \"lib/**/*.tsx\" -e \"lib/*/node_modules/**\" -e \"lib/*/dist/**\"",
-    "postinstall": "npm run build:action-library",
-    "build:action-library": "cd lib/action-library && rm -rf dist && tsc",
-    "build:ui": "NODE_ENV=production webpack",
-    "dev:ui": "NODE_ENV=dev API_URL='http://localhost:8000/' webpack-dev-server --color",
-    "heroku-postbuild": "npm run build:ui",
-    "versionist": "versionist",
-    "addgittag": "./scripts/add-version-tag.js"
+    "postinstall": "make build",
+    "heroku-postbuild": "make build"
   },
   "ava": {
     "timeout": "5m",
@@ -91,7 +78,6 @@
     "webpack-dev-server": "^3.1.1"
   },
   "dependencies": {
-    "@resin.io/jellyfish-action-library": "file:./lib/action-library",
     "@resin.io/jellyfish-sdk": "file:./lib/sdk",
     "@types/blueimp-md5": "^2.7.0",
     "@types/es6-promise": "^3.3.0",
@@ -100,6 +86,7 @@
     "@types/json-schema": "^6.0.1",
     "@types/lodash": "^4.14.104",
     "@types/mark.js": "^8.11.1",
+    "@types/node": "^10.9.4",
     "@types/react": "^16.0.40",
     "@types/react-dom": "^16.0.4",
     "@types/react-redux": "^5.0.19",
@@ -143,7 +130,7 @@
     "rendition": "^4.18.1",
     "reql-schema": "^1.0.0",
     "rethinkdb": "^2.3.3",
-    "skhema": "^2.1.1",
+    "skhema": "^2.2.0",
     "socket.io": "^2.1.0",
     "static-eval": "^2.0.0",
     "styled-components": "^3.2.5",
