+++ test/permissions.spec.js
@@ -24,7 +24,7 @@ const randomstring = require('randomstring')
 const {
 	getSdk
 } = require('@resin.io/jellyfish-sdk')
-const createServer = require('../lib/server')
+const createServer = require('../lib/server/create-server')
 
 const WAIT_TIMEOUT = 30 * 1000
